자신의 글만 삭제 가능

+++ src/main/java/com/example/prjjsp2/service/BoardService.java
@@ -61,8 +61,13 @@ public class BoardService {
         return board;
     }
 
-    public void remove(Integer id) {
-        mapper.deleteById(id);
+    public void remove(Integer id, Member member) {
+        Board board = mapper.selectById(id);
+        if (board.getWriter().equals(member.getId())) {
+            mapper.deleteById(id);
+        } else {
+            throw new RuntimeException("삭제 권한이 없습니다.");
+        }
     }
 
     public void update(Board board) {
